refactor(kicad9): relocate label-deconfliction decision to schematics/
Move the overlap-detection + nudge-target decision of _deconflict_labels into
schematics/decisions.deconflict_labels, returning abstract (idx, nx, ny) moves
and (ax, ay, nx, ny) connecting-wire segments in render-mm. The Sexp reading,
 mutation and dcwire construction stay in the backend
(Kicad9Backend.apply_label_deconfliction). Occupancy seed is built in element
order and passed in so later entries overwrite earlier ones at a shared cell,
preserving the original single-pass semantics. Verified byte-identical to the
old _deconflict_labels across 200 randomized synthetic element lists.

(cherry picked from commit 513989bd7b1ae8bf4258c742b8839bbdb52f6966)
